requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
blinker==1.8.2
boto3==1.28.23
botocore==1.31.54
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
clip @ git+https://github.com/ultralytics/CLIP.git@e17416a36b45d040758327936a1ea150c13fe3d1
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.7
cycler==0.10.0
Cython==3.0.0
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.1
fastapi==0.110.3
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
GPUtil==1.4.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==2.10
imageio==2.34.1
importlib_resources==6.4.0
inference==0.10.0
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
mypy-extensions==1.0.0
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
onnxruntime==1.15.1
openai==1.30.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
pluggy==1.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydot==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytest==8.2.0
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==306
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
redis==5.0.4
referencing==0.35.1
regex==2024.5.15
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.29
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.6.2
scikit-image==0.22.0
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.34.0
structlog==24.1.0
supervision==0.20.0
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
thop==0.1.1.post2209072238
tifffile==2024.5.10
time-machine==2.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typer==0.9.0
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.17
uritemplate==4.1.1
urllib3==1.26.18
watchdog==4.0.0
wcwidth==0.2.13
websocket-client==1.8.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
yarl==1.9.4
zipp==3.18.2
zxing-cpp==2.2.0
